+++ wgpu-types/src/lib.rs
@@ -71,11 +71,32 @@ pub const QUERY_SET_MAX_QUERIES: u32 = 4096;
 pub const QUERY_SIZE: u32 = 8;
 
 /// Backends supported by wgpu.
+///
+/// See also [`Backends`].
 #[repr(u8)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub enum Backend {
     /// Dummy backend, which may be used for testing.
+    ///
+    /// It performs no rendering or computation, but allows creation of stub GPU resource types,
+    /// so that code which manages GPU resources can be tested without an available GPU.
+    /// Specifically, the following operations are implemented:
+    ///
+    /// * Enumerating adapters will always return one noop adapter, which can be used to create
+    ///   devices.
+    /// * Buffers may be created, written, mapped, and copied to other buffers.
+    /// * Command encoders may be created, but only buffer operations are useful.
+    ///
+    /// Other resources can be created but are nonfunctional; notably,
+    ///
+    /// * Render passes and compute passes are not executed.
+    /// * Textures may be created, but do not store any texels.
+    /// * There are no compatible surfaces.
+    ///
+    /// An adapter using the noop backend can only be obtained if [`NoopBackendOptions`]
+    /// enables it, in addition to the ordinary requirement of [`Backends::NOOP`] being set.
+    /// This ensures that applications not desiring a non-functional backend will not receive it.
     Noop = 0,
     /// Vulkan API (Windows, Linux, Android, MacOS via `vulkan-portability`/MoltenVK)
     Vulkan = 1,
@@ -148,22 +169,35 @@ bitflags::bitflags! {
     #[cfg_attr(feature = "serde", serde(transparent))]
     #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
     pub struct Backends: u32 {
+        /// [`Backend::Noop`].
+        const NOOP = 1 << Backend::Noop as u32;
+
+        /// [`Backend::Vulkan`].
         /// Supported on Windows, Linux/Android, and macOS/iOS via Vulkan Portability (with the Vulkan feature enabled)
         const VULKAN = 1 << Backend::Vulkan as u32;
+
+        /// [`Backend::Gl`].
         /// Supported on Linux/Android, the web through webassembly via WebGL, and Windows and
         /// macOS/iOS via ANGLE
         const GL = 1 << Backend::Gl as u32;
-        /// Supported on macOS/iOS
+
+        /// [`Backend::Metal`].
+        /// Supported on macOS and iOS.
         const METAL = 1 << Backend::Metal as u32;
+
+        /// [`Backend::Dx12`].
         /// Supported on Windows 10 and later
         const DX12 = 1 << Backend::Dx12 as u32;
-        /// Supported when targeting the web through webassembly with the `webgpu` feature enabled.
+
+        /// [`Backend::BrowserWebGpu`].
+        /// Supported when targeting the web through WebAssembly with the `webgpu` feature enabled.
         ///
         /// The WebGPU backend is special in several ways:
         /// It is not not implemented by `wgpu_core` and instead by the higher level `wgpu` crate.
         /// Whether WebGPU is targeted is decided upon the creation of the `wgpu::Instance`,
         /// *not* upon adapter creation. See `wgpu::Instance::new`.
         const BROWSER_WEBGPU = 1 << Backend::BrowserWebGpu as u32;
+
         /// All the apis that wgpu offers first tier of support for.
         ///
         /// * [`Backends::VULKAN`]
@@ -174,6 +208,7 @@ bitflags::bitflags! {
             | Self::METAL.bits()
             | Self::DX12.bits()
             | Self::BROWSER_WEBGPU.bits();
+
         /// All the apis that wgpu offers second tier of support for. These may
         /// be unsupported/still experimental.
         ///
@@ -233,6 +268,7 @@ impl Backends {
                 "metal" | "mtl" => Self::METAL,
                 "opengl" | "gles" | "gl" => Self::GL,
                 "webgpu" => Self::BROWSER_WEBGPU,
+                "noop" => Self::NOOP,
                 b => {
                     log::warn!("unknown backend string '{}'", b);
                     continue;